LM-683: SAML - Rename laa-saml-mock project and artifacts
Renaming the groupId and version of the pom modules, and the .jar
artifacts in the sp/idp target directories.

Instructions are available in the readme.md file on how to
start the spring boot applications along with convenience
start-idp.sh/start-sp.sh script files, and how to install
the Mock SAML services in an AWS EC2 instance.