Do not send SubjectConfirmationData Address= with URL
Before, Mujina would send the following SubjectConfirmationData:
```xml
 <saml2:SubjectConfirmationData Address="https://engine.example.org/authentication/sp/consume-assertion"
```

The Address attribute is in SAML 2.0 core more or less defined to be
an IP-address (or at least something a session can be "bound" to).
This being an URL trips some very strict libraries.

IP-address binding is not a feature that is useful in Mujina (or in
general, if you ask me). Likely best just not to send it at all.