Fix the add attribute select in ie 10/11
Prior to this change, the select did not work because the event handler used the selectedOptions property, which does not work in IE (see: https://caniuse.com/?search=selectedoptions).

This change reworks the code to use the value and select the correct option from there.  All of that code is compatible with IE10/11 and modern browsers.